refactor: opt code and comments
