+++ packages/core/src/index.ts
@@ -11,7 +11,7 @@ export { Component } from "./Component";
 export { Script } from "./Script";
 export { Renderer } from "./Renderer";
 export { dependentComponents, DependentMode } from "./ComponentsDependencies";
-export { Camera, MSAASamples } from "./Camera";
+export { Camera } from "./Camera";
 export { Transform } from "./Transform";
 export { BoolUpdateFlag } from "./BoolUpdateFlag";
 export type { EngineSettings } from "./EngineSettings";
@@ -35,6 +35,8 @@ export { DepthTextureMode } from "./enums/DepthTextureMode";
 export { FogMode } from "./enums/FogMode";
 export { CameraClearFlags } from "./enums/CameraClearFlags";
 export { CameraType } from "./enums/CameraType";
+export { MSAASamples } from "./enums/MSAASamples";
+export { Downsampling } from "./enums/Downsampling";
 export { ColorSpace } from "./enums/ColorSpace";
 export { BackgroundTextureFillMode } from "./enums/BackgroundTextureFillMode";
 export { XRManager } from "./xr/XRManager";
